REFACTOR: Move typeahead initialisation code into own module
This helps to keep the index file clean by moving all the code to setup the typeahead elements into its own module.
